Migrate message completeness test to new test suite
also remove test files no longer in use
