+++ docs-xml/mkmsgs.py
@@ -1,10 +1,13 @@
 import sys
+from os.path import isfile
 import os.path as path
 basename = path.dirname(path.dirname(path.abspath(__file__)))
 sys.path.append(path.join(basename,'src'))
 
 from feedvalidator.i18n.en import messages
 from feedvalidator.logging import Warning, Error
+import feedvalidator
+
 
 template = '''
 <fvdoc>
@@ -22,8 +25,42 @@ template = '''
 </fvdoc>
 '''
 
+import inspect
+# Logic from text_html.py
+def getRootClass(aClass):
+  bl = aClass.__bases__
+  if not(bl):
+    return None
+
+  aClass = bl[0]
+  bl = bl[0].__bases__
+
+  while bl:
+    base = bl[0]
+    if base == feedvalidator.logging.Message:
+      return aClass
+    aClass = base
+    bl = aClass.__bases__
+  return None
+
+def isclass(x):
+  import types
+  return inspect.isclass(x) or type(x) == type
+
 def missing():
   result = []
+
+  for n, o in inspect.getmembers(feedvalidator.logging, isclass):
+    rc = getRootClass(o)
+    if not(rc):
+      continue
+
+    rcname = rc.__name__.split('.')[-1].lower()
+    if rcname in ['warning', 'error']:
+      fn = path.join(basename, 'docs', rcname, n + '.html')
+      if not(isfile(path.join(basename, fn))):
+        result.append((rcname, n, "", fn, fn))
+
 
   for key, value in list(messages.items()):
     if issubclass(key,Error):
@@ -33,10 +70,11 @@ def missing():
     else:
       continue
 
+    html = path.join(basename, 'docs', dir, key.__name__+'.html')
     xml = path.join(basename, 'docs-xml', dir, key.__name__+'.xml')
 
-    if not path.exists(xml):
-      result.append((dir, key.__name__, value, xml))
+    if not path.exists(xml) or not path.exists(html):
+      result.append((dir, key.__name__, value, html, xml))
 
   return result
 
@@ -44,7 +82,7 @@ import unittest
 class MissingMessagesTest(unittest.TestCase):
   def test_messages(self):
     self.assertEqual([],
-      ["%s/%s" % (dir,id) for dir, id, msg, xml, html in missing()])
+                     ["%s/%s" % (dir,id) for dir, id, msg, xml, html in missing()], "Errors/warnings without corresponding documentation")
 
 def buildTestSuite():
   suite = unittest.TestSuite()
